+++ packages/amplify-category-auth/src/provider-utils/awscloudformation/assets/cognito-defaults.js
@@ -96,4 +96,5 @@ module.exports = {
   withSocialDefaults,
   entityKeys,
   roles,
+  sharedId,
 };