+++ packages/amplify-category-function/src/provider-utils/awscloudformation/utils/build.ts
@@ -1,11 +1,15 @@
 import { $TSContext } from 'amplify-cli-core';
 import { buildFunction, BuildRequestMeta } from './buildFunction';
+import { buildLayer } from './buildLayer';
 import { ServiceName } from './constants';
 
 export const buildResource = (context: $TSContext, resource: BuildRequestMeta & { service: string }) => {
-  // only build lambda functions
-  if (resource.service !== ServiceName.LambdaFunction) {
+  // only build lambda functions and lambda layers
+  if (resource.service !== ServiceName.LambdaFunction && resource.service !== ServiceName.LambdaLayer) {
     return;
   }
-  return buildFunction(context, resource);
+  return getBuilderForService(resource.service)(context, resource);
 };
+
+// since lambdaLayers build is dfifferent from lambda function
+const getBuilderForService = (service: string) => (service === ServiceName.LambdaLayer ? buildLayer : buildFunction);