+++ packages/amplify-python-function-runtime-provider/src/constants.ts
@@ -1,4 +1,5 @@
-import { join } from 'path';
+import * as path from 'path';
 
 export const packageName = 'amplify-python-function-runtime-provider';
-export const relativeShimPath = join('shim', 'shim.py');
+export const relativeShimPath = path.join('shim', 'shim.py');
+export const layerPythonPipFile = path.join(__dirname, '..', 'resources', 'Pipfile');