+++ packages/amplify-util-import/CHANGELOG.md
@@ -3,6 +3,37 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.5.1](https://github.com/aws-amplify/amplify-cli/compare/amplify-util-import@1.5.0...amplify-util-import@1.5.1) (2021-06-15)
+
+**Note:** Version bump only for package amplify-util-import
+
+
+
+
+
+# [1.5.0](https://github.com/aws-amplify/amplify-cli/compare/amplify-util-import@1.4.9...amplify-util-import@1.5.0) (2021-06-02)
+
+
+
+# 4.52.0 (2021-06-01)
+
+
+### Features
+
+* add support for SMS Sandbox ([#7436](https://github.com/aws-amplify/amplify-cli/issues/7436)) ([cdcb626](https://github.com/aws-amplify/amplify-cli/commit/cdcb6260c11bbedef5b056fdcd730612d8bb3230))
+
+
+
+
+
+## [1.4.9](https://github.com/aws-amplify/amplify-cli/compare/amplify-util-import@1.4.8...amplify-util-import@1.4.9) (2021-05-26)
+
+**Note:** Version bump only for package amplify-util-import
+
+
+
+
+
 ## [1.4.8](https://github.com/aws-amplify/amplify-cli/compare/amplify-util-import@1.4.7...amplify-util-import@1.4.8) (2021-05-18)
 
 **Note:** Version bump only for package amplify-util-import